requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
